Migrate Security Group ids validator
